pyproject.toml
[project]
name = "arxiv-rag-copilot"
version = "0.3.0"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=0.4.22",
    "datasets>=4.5.0",
    "fastapi>=0.128.0",
    "feedparser>=6.0.12",
    "httpx>=0.28.1",
    "jinja2>=3.1.0",
    "langchain>=0.3.0",
    "langchain-chroma>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-huggingface>=0.1.0",
    "langchain-openai>=0.2.0",
    "langchain-postgres>=0.0.12",
    "nltk>=3.9.2",
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "pgvector>=0.3.0",
    "prometheus-client>=0.19.0",
    "psycopg>=3.2.0",
    "psycopg[binary]>=3.2.0",
    "pydantic-settings>=2.12.0",
    "ragas>=0.4.3",
    "rank-bm25>=0.2.2",
    "redis>=5.0.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.2",
    "spacy>=3.7.0",
    "sse-starlette>=3.2.0",
    "structlog>=25.5.0",
    "tenacity>=9.0.0",
    "sqlalchemy>=2.0.0",
    "uvicorn>=0.32.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-httpx>=0.60b1",
    "opentelemetry-exporter-otlp-proto-grpc>=1.39.1",
    "opentelemetry-exporter-jaeger-thrift>=1.21.0",
    # Pydantic for data validation
    "pydantic>=2.0.0",
    # Security & Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "authlib>=1.2.1",
    # Database (async driver)
    "asyncpg>=0.29.0",
    "alembic>=1.12.0",
    # Performance
    "tqdm>=4.66.0",
    # Email validation
    "email-validator>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "coverage[toml]>=7.6.0",
    "httpx>=0.28.0",
    "fakeredis>=2.25.0",
    # Load testing
    "locust>=2.0.0",
    "pytest-benchmark>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=70",
]
markers = [
    "asyncio: mark test as async",
    "integration: mark test as integration test",
    "slow: mark test as slow running",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
    "*/venv/*",
    "*/.venv/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod",
]
precision = 2
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"
